Notifications for ADD/EDIT 17-07-2022

+++ Tasks/src/main/java/com/capstone/Tasks/Entity/Tasks.java
@@ -13,7 +13,7 @@ import java.util.List;
 public class Tasks
 {
     @Id
-    private int taskId;
+    private long taskId;
     private String taskName;
     private String taskDescription;
     private String priority;
